// src/lib/dawg-types.ts
export interface LocationItem {
  id: string;
  name: string;
  type: string;
  address: string;
  cityStateZip: string;
  phone: string;
  email: string;
  manager: string;
  status: 'Active' | 'Under Maintenance' | 'Opening Soon';
  stationCount: number;
  operatingHours: string;
  isDefault: boolean;
}